Fix struct cast operator
Previously it was using the wrong API, which led to using the void return
value for the struct and wrong arguments.

LLVM needs an attribute set to specify that an argument is going to written
back to with a struct. It works on some archs without this but use with all
archs. This is added to struct message sends.

+ Set target tripple on module
+ Don't pass arguments to main wrapper from JS

This fixes #9!!!!!

Known issues: objc_Struct cast only works with a single struct value access
i.e. anObject.value. In the near future casting will be handled
by the compiler and this will be addressed then.